拆分： 1.将WeiXinInterface接口基础化; 2.将WeiXinService服务基础化; 3.分离支付业务类型;
